removing unused mcmc_gw.h functions and moving used ones to mcmc_gw_extended
