+++ apps/CMakeLists.txt
@@ -16,14 +16,6 @@ target_include_directories( create_gw_init_file PUBLIC "${INCLUDE_DIRS}")
 #target_include_directories( create_gw_init_file PUBLIC "${PROJECT_SOURCE_DIR}/include")
 #target_include_directories( create_gw_init_file PUBLIC "${OpenMP_INCLUDE_DIR}" "${ADOLC_INCLUDE_DIR}" "${HDF5_INCLUDE_DIR}" "${FFTW3_INCLUDE_DIR}" "${GSL_INCLUDE_DIR}" "${PTRJMCMC_INCLUDE_DIR}")
 
-add_executable(mcmc_gw_tool "mcmc_gw_tool.cpp")
-set_target_properties(mcmc_gw_tool PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${PROJECT_BINARY_DIR}/bin")
-target_link_libraries(mcmc_gw_tool PUBLIC gwat)
-target_link_libraries( mcmc_gw_tool PUBLIC "${LIB_DIRS}")
-target_include_directories( mcmc_gw_tool PUBLIC "${INCLUDE_DIRS}")
-#target_link_libraries( mcmc_gw_tool PUBLIC "${HDF5_LIBRARIES}" OpenMP::OpenMP_CXX "${ADOLC_LIBRARIES}" "${FFTW3_LIBRARIES}" "${GSL_LIBRARIES}" "${PTRJMCMC_LIBRARIES}")
-#target_include_directories( mcmc_gw_tool PUBLIC "${PROJECT_SOURCE_DIR}/include")
-#target_include_directories( mcmc_gw_tool PUBLIC "${OpenMP_INCLUDE_DIR}" "${ADOLC_INCLUDE_DIR}" "${HDF5_INCLUDE_DIR}" "${FFTW3_INCLUDE_DIR}" "${GSL_INCLUDE_DIR}" "${Eigen3_INCLUDE_DIR}" "${PTRJMCMC_INCLUDE_DIR}")
 
 add_executable(mcmc_gw_tool_v2 "mcmc_gw_tool_v2.cpp")
 set_target_properties(mcmc_gw_tool_v2 PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${PROJECT_BINARY_DIR}/bin")
@@ -31,16 +23,6 @@ target_link_libraries(mcmc_gw_tool_v2 PUBLIC gwat)
 target_link_libraries( mcmc_gw_tool_v2 PUBLIC "${LIB_DIRS}")
 target_include_directories( mcmc_gw_tool_v2 PUBLIC "${INCLUDE_DIRS}")
 
-
-
-add_executable(rjmcmc_gw_tool_2wf "rjmcmc_gw_tool_2wf.cpp")
-set_target_properties(rjmcmc_gw_tool_2wf PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${PROJECT_BINARY_DIR}/bin")
-target_link_libraries(rjmcmc_gw_tool_2wf PUBLIC gwat)
-target_link_libraries( rjmcmc_gw_tool_2wf PUBLIC "${LIB_DIRS}")
-target_include_directories( rjmcmc_gw_tool_2wf PUBLIC "${INCLUDE_DIRS}")
-#target_link_libraries( rjmcmc_gw_tool_2wf PUBLIC "${HDF5_LIBRARIES}" OpenMP::OpenMP_CXX "${ADOLC_LIBRARIES}" "${FFTW3_LIBRARIES}" "${GSL_LIBRARIES}" "${PTRJMCMC_LIBRARIES}")
-#target_include_directories( rjmcmc_gw_tool_2wf PUBLIC "${PROJECT_SOURCE_DIR}/include")
-#target_include_directories( rjmcmc_gw_tool_2wf PUBLIC "${OpenMP_INCLUDE_DIR}" "${ADOLC_INCLUDE_DIR}" "${HDF5_INCLUDE_DIR}" "${FFTW3_INCLUDE_DIR}" "${GSL_INCLUDE_DIR}" "${PTRJMCMC_INCLUDE_DIR}")
 
 add_executable(fisher_tool "fisher_tool.cpp")
 set_target_properties(fisher_tool PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${PROJECT_BINARY_DIR}/bin")
@@ -61,9 +43,7 @@ target_include_directories( autocorr_tool PUBLIC "${INCLUDE_DIRS}")
 #target_include_directories( autocorr_tool PUBLIC "${OpenMP_INCLUDE_DIR}" "${ADOLC_INCLUDE_DIR}" "${HDF5_INCLUDE_DIR}" "${FFTW3_INCLUDE_DIR}" "${GSL_INCLUDE_DIR}" "${PTRJMCMC_INCLUDE_DIR}")
 
 install(TARGETS waveform_tool DESTINATION bin)
-install(TARGETS mcmc_gw_tool DESTINATION bin)
 install(TARGETS mcmc_gw_tool_v2 DESTINATION bin)
-install(TARGETS rjmcmc_gw_tool_2wf DESTINATION bin)
 install(TARGETS autocorr_tool DESTINATION bin)
 install(TARGETS fisher_tool DESTINATION bin)
 install(TARGETS create_gw_init_file DESTINATION bin)
