--- gwatpy/gwatpy/config.py
@@ -1,5 +1,0 @@
-#System specific configuration file for gwatpy
-#LIB="/home/sperkins/git-repos/gw_analysis_tools/build/install/lib/libgwat.so"
-LIB="/Users/kristenschumacher/Desktop/Gravity/Coding/gw_analysis_tools/build/install/lib/libgwat.dylib"
-#GWATPY_ROOT_DIRECTORY="/home/sperkins/git-repos/gw_analysis_tools/gwatpy/gwatpy/"
-GWATPY_ROOT_DIRECTORY="/Users/kristenschumacher/Desktop/Gravity/Coding/gw_analysis_tools/gwatpy/gwatpy/"
