+++ include/gwat/mcmc_gw.h
@@ -140,502 +140,5 @@ struct MCMC_user_param
 };
 
 
-double maximized_coal_log_likelihood_IMRPhenomD(double *frequencies,
-				int length,
-				std::complex<double> *data,
-				double *noise,
-				double SNR,
-				double chirpmass,	
-				double symmetric_mass_ratio, 
-				double spin1,
-				double spin2,
-				bool NSflag,
-				fftw_outline *plan);
-double maximized_coal_log_likelihood_IMRPhenomD(double *frequencies,
-				size_t length,
-				double *real_data,
-				double *imag_data,
-				double *noise,
-				double SNR,
-				double chirpmass,	
-				double symmetric_mass_ratio, 
-				double spin1,
-				double spin2,
-				bool NSflag);
-
-double maximized_coal_log_likelihood_IMRPhenomD(double *frequencies,
-				size_t length,
-				double *real_data,
-				double *imag_data,
-				double *noise,
-				double SNR,
-				double chirpmass,	
-				double symmetric_mass_ratio, 
-				double spin1,
-				double spin2,
-				bool NSflag,
-				fftw_outline *plan);
-double maximized_coal_log_likelihood_IMRPhenomD_Full_Param(double *frequencies,
-				int length,
-				std::complex<double> *data,
-				double *noise,
-				double chirpmass,	/**< in solar masses*/
-				double symmetric_mass_ratio, 
-				double spin1,
-				double spin2,
-				double Luminosity_Distance,
-				double theta,
-				double phi,
-				double iota,
-				bool NSflag,
-				fftw_outline *plan);
-
-double maximized_coal_log_likelihood_IMRPhenomD_Full_Param(double *frequencies,
-				size_t length,
-				double *real_data,
-				double *imag_data,
-				double *noise,
-				double chirpmass,	/**< in solar masses*/
-				double symmetric_mass_ratio, 
-				double spin1,
-				double spin2,
-				double Luminosity_Distance,
-				double theta,
-				double phi,
-				double iota,
-				bool NSflag);
-
-double maximized_coal_log_likelihood_IMRPhenomD_Full_Param(double *frequencies,
-				size_t length,
-				double *real_data,
-				double *imag_data,
-				double *noise,
-				double chirpmass,	/**< in solar masses*/
-				double symmetric_mass_ratio, 
-				double spin1,
-				double spin2,
-				double Luminosity_Distance,
-				double theta,
-				double phi,
-				double iota,
-				bool NSflag,
-				fftw_outline *plan);
-
-
-double maximized_Log_Likelihood_aligned_spin_internal(std::complex<double> *data,
-				double *psd,
-				double *frequencies,
-				std::complex<double> *detector_response,
-				size_t length,
-				fftw_outline *plan
-				);
-double Log_Likelihood(std::complex<double> *data,
-				double *psd,
-				double *frequencies,
-				size_t length,
-				gen_params_base<double> *params,
-				std::string detector,
-				std::string generation_method
-				);
-
-double maximized_Log_Likelihood_unaligned_spin_internal(std::complex<double> *data,
-				double *psd,
-				double *frequencies,
-				std::complex<double> *hplus,
-				std::complex<double> *hcross,
-				size_t length,
-				fftw_outline *plan
-				);
-double maximized_Log_Likelihood(std::complex<double> *data, 
-				double *psd,
-				double *frequencies,
-				size_t length,
-				gen_params_base<double> *params,
-				std::string detector,
-				std::string generation_method,
-				fftw_outline *plan
-				);
-double maximized_Log_Likelihood(double *data_real, 
-				double *data_imag,
-				double *psd,
-				double *frequencies,
-				size_t length,
-				//gen_params *params,
-				gen_params_base<double> *params,
-				std::string detector,
-				std::string generation_method,
-				fftw_outline *plan
-				);
-double maximized_coal_Log_Likelihood(double *data_real, 
-				double *data_imag,
-				double *psd,
-				double *frequencies,
-				size_t length,
-				double *template_real,
-				double *template_imag,
-				fftw_outline *plan
-				);
-double maximized_coal_Log_Likelihood(std::complex<double> *data, 
-				double *psd,
-				double *frequencies,
-				size_t length,
-				gen_params_base<double> *params,
-				std::string detector,
-				std::string generation_method,
-				fftw_outline *plan,
-				double *tc,
-				double *phic
-				);
-double maximized_coal_Log_Likelihood_internal(std::complex<double> *data,
-				double *psd,
-				double *frequencies,
-				std::complex<double> *detector_response,
-				size_t length,
-				fftw_outline *plan,
-				double *tc,
-				double *phic
-				);
-double Log_Likelihood_internal(std::complex<double> *data,
-			double *psd,
-			double *frequencies,
-			double *weights,
-			std::complex<double> *detector_response,
-			int length,
-			bool log10F,
-			std::string integration_method
-			);
-double Log_Likelihood_internal(std::complex<double> *data,
-	double *psd,
-	std::complex<double> *detector_response,
-	Quadrature *QuadMethod
-);
-
-double MCMC_likelihood_wrapper_SKYSEARCH(double *param, mcmc_data_interface *interface,void *parameters);
-void SkySearch_PTMCMC_MH_dynamic_PT_alloc_uncorrelated_GW(mcmc_sampler_output *sampler_output,
-	double **output,
-	int dimension,
-	int N_steps,
-	int chain_N,
-	int max_chain_N_thermo_ensemble,
-	double *initial_pos,
-	double *seeding_var,
-	double **ensemble_initial_pos,
-	double *chain_temps,
-	int swp_freq,
-	int t0,
-	int nu,
-	int max_chunk_size,
-	std::string chain_distribution_scheme,
-	double(*log_prior)(double *param, mcmc_data_interface *interface,void *parameters),
-	int numThreads,
-	bool pool,
-	bool show_prog,
-	int num_detectors,
-	std::complex<double> **data,
-	double **noise_psd,
-	double **frequencies,
-	int *data_length,
-	double gps_time,
-	std::string *detectors,
-	int Nmod,
-	double *bppe,
-	std::complex<double> *hplus,
-	std::complex<double> *hcross,
-	std::string statistics_filename,
-	std::string chain_filename,
-	std::string likelihood_log_filename,
-	std::string checkpoint_filename
-	);
-
-void PTMCMC_MH_GW(double ***output,
-	int dimension,
-	int N_steps,
-	int chain_N,
-	double *initial_pos,
-	double *seeding_var,
-	double **ensemble_initial_pos,
-	double *chain_temps,
-	int swp_freq,
-	double(*log_prior)(double *param, mcmc_data_interface *interface,void *parameters),
-	int numThreads,
-	bool pool,
-	bool show_prog,
-	int num_detectors,
-	std::complex<double> **data,
-	double **noise_psd,
-	double **frequencies,
-	int *data_length,
-	double gps_time,
-	std::string *detector,
-	MCMC_modification_struct *mod_struct,
-	std::string generation_method,
-	std::string statistics_filename,
-	std::string chain_filename,
-	std::string auto_corr_filename,
-	std::string likelihood_log_filename,
-	std::string checkpoint_filename
-	);
-void continue_PTMCMC_MH_dynamic_PT_alloc_uncorrelated_GW(std::string checkpoint_file_start,
-	mcmc_sampler_output *sampler_output,
-	double **output,
-	int N_steps,
-	int max_chain_N_thermo_ensemble,
-	double *chain_temps,
-	int swp_freq,
-	int t0,
-	int nu,
-	int max_chunk_size,
-	std::string chain_distribution_scheme,
-	double(*log_prior)(double *param, mcmc_data_interface *interface,void *parameters),
-	int numThreads,
-	bool pool,
-	bool show_prog,
-	int num_detectors,
-	std::complex<double> **data,
-	double **noise_psd,
-	double **frequencies,
-	int *data_length,
-	double gps_time,
-	std::string *detectors,
-	MCMC_modification_struct *mod_struct,
-	std::string generation_method,
-	std::string statistics_filename,
-	std::string chain_filename,
-	std::string likelihood_log_filename,
-	std::string checkpoint_filename
-	);
-void PTMCMC_MH_dynamic_PT_alloc_uncorrelated_GW(mcmc_sampler_output *sampler_output,
-	double **output,
-	int dimension,
-	int N_steps,
-	int chain_N,
-	int max_chain_N_thermo_ensemble,
-	double *initial_pos,
-	double *seeding_var,
-	double **ensemble_initial_pos,
-	double *chain_temps,
-	int swp_freq,
-	int t0,
-	int nu,
-	int max_chunk_size,
-	std::string chain_distribution_scheme,
-	double(*log_prior)(double *param, mcmc_data_interface *interface,void *parameters),
-	int numThreads,
-	bool pool,
-	bool show_prog,
-	int num_detectors,
-	std::complex<double> **data,
-	double **noise_psd,
-	double **frequencies,
-	int *data_length,
-	double gps_time,
-	std::string *detectors,
-	MCMC_modification_struct *mod_struct,
-	std::string generation_method,
-	std::string statistics_filename,
-	std::string chain_filename,
-	std::string likelihood_log_filename,
-	std::string checkpoint_filename
-	);
-void PTMCMC_MH_dynamic_PT_alloc_GW(double ***output,
-	int dimension,
-	int N_steps,
-	int chain_N,
-	int max_chain_N_thermo_ensemble,
-	double *initial_pos,
-	double *seeding_var,
-	double **ensemble_initial_pos,
-	double *chain_temps,
-	int swp_freq,
-	int t0,
-	int nu,
-	std::string chain_distribution_scheme,
-	double(*log_prior)(double *param, mcmc_data_interface *interface ,void *parameters),
-	int numThreads,
-	bool pool,
-	bool show_prog,
-	int num_detectors,
-	std::complex<double> **data,
-	double **noise_psd,
-	double **frequencies,
-	int *data_length,
-	double gps_time,
-	std::string *detectors,
-	MCMC_modification_struct *mod_struct,
-	std::string generation_method,
-	std::string statistics_filename,
-	std::string chain_filename,
-	std::string likelihood_log_filename,
-	std::string checkpoint_filename
-	);
-void continue_PTMCMC_MH_GW(std::string start_checkpoint_file,
-	double ***output,
-	int dimension,
-	int N_steps,
-	int swp_freq,
-	double(*log_prior)(double *param, mcmc_data_interface *interface,void *parameters),
-	int numThreads,
-	bool pool,
-	bool show_prog,
-	int num_detectors,
-	std::complex<double> **data,
-	double **noise_psd,
-	double **frequencies,
-	int *data_length,
-	double gps_time,
-	std::string *detector,
-	MCMC_modification_struct *mod_struct,
-	std::string generation_method,
-	std::string statistics_filename,
-	std::string chain_filename,
-	std::string auto_corr_filename,
-	std::string likelihood_log_filename,
-	std::string final_checkpoint_filename,
-	bool tune
-	);
-void PTMCMC_method_specific_prep(std::string generation_method, int dimension,double **seeding_var_ptr, bool local_seeding);
-
-void RJPTMCMC_method_specific_prep(std::string generation_method, int max_dim, int min_dim,double *seeding_var, bool local_seeding);
-
-double MCMC_likelihood_extrinsic(bool save_waveform, 
-	gen_params_base<double> *parameters,
-	std::string generation_method, 
-	int *data_length, 
-	double **frequencies, 
-	std::complex<double> **data, 
-	double **psd,  
-	double **weights, 
-	std::string integration_method, 
-	bool log10F,
-	std::string *detectors, 
-	int num_detectors,
-	Quadrature *QuadMethod = NULL
-	);
-
-void MCMC_fisher_wrapper(double *param,  double **output, mcmc_data_interface *interface ,void *parameters) ;
-
-std::string MCMC_prep_params(double *param, double *temp_params, gen_params_base<double> *gen_params, int dimension, std::string generation_method,MCMC_modification_struct *mod_struct);
-
-double MCMC_likelihood_wrapper(double *param, mcmc_data_interface *interface ,void *parameters) ;
-
-void continue_RJPTMCMC_MH_dynamic_PT_alloc_comprehensive_2WF_GW(
-	std::string checkpoint_file_start,
-	mcmc_sampler_output *sampler_output,
-	double **output,
-	int **status,
-	int *model_status,
-	int nested_model_number,
-	int N_steps,
-	int max_chain_N_thermo_ensemble,
-	double **prior_ranges,
-	double *chain_temps,
-	int swp_freq,
-	int t0,
-	int nu,
-	int max_chunk_size,
-	std::string chain_distribution_scheme,
-	double(*log_prior)(double *param, int *status,int model_status,mcmc_data_interface *interface,void *parameters),
-	int numThreads,
-	bool pool,
-	bool show_prog,
-	int num_detectors,
-	std::complex<double> **data,
-	double **noise_psd,
-	double **frequencies,
-	int *data_length,
-	double gps_time,
-	std::string *detectors,
-	MCMC_modification_struct *mod_struct,
-	std::string generation_method_base,
-	std::string generation_method_extended,
-	std::string statistics_filename,
-	std::string chain_filename,
-	std::string likelihood_log_filename,
-	std::string checkpoint_filename
-	);
-void RJPTMCMC_MH_dynamic_PT_alloc_comprehensive_2WF_GW(
-	mcmc_sampler_output *sampler_output,
-	double **output,
-	int **status,
-	int *model_status,
-	int nested_model_number,
-	int max_dimension,
-	int min_dimension,
-	int N_steps,
-	int chain_N,
-	int max_chain_N_thermo_ensemble,
-	double *initial_pos,
-	int *initial_status,
-	int initial_model_status,
-	double *seeding_var,
-	double **ensemble_initial_pos,
-	int **ensemble_initial_status,
-	int *ensemble_initial_model_status,
-	double **prior_ranges,
-	double *chain_temps,
-	int swp_freq,
-	int t0,
-	int nu,
-	int max_chunk_size,
-	std::string chain_distribution_scheme,
-	double(*log_prior)(double *param, int *status,int model_status,mcmc_data_interface *interface,void *parameters),
-	int numThreads,
-	bool pool,
-	bool show_prog,
-	int num_detectors,
-	std::complex<double> **data,
-	double **noise_psd,
-	double **frequencies,
-	int *data_length,
-	double gps_time,
-	std::string *detectors,
-	MCMC_modification_struct *mod_struct,
-	std::string generation_method_base,
-	std::string generation_method_extended,
-	std::string statistics_filename,
-	std::string chain_filename,
-	std::string likelihood_log_filename,
-	std::string checkpoint_filename
-	);
-void RJMCMC_2WF_RJ_proposal_wrapper(
-	double *current_param, 
-	double *proposed_param, 
-	int *current_status, 
-	int *proposed_status, 
-	int *current_model_status, 
-	int *proposed_model_status, 
-	double *MH_corrections,
-	mcmc_data_interface *interface, 
-	void *user_param);
-void RJMCMC_2WF_fisher_wrapper(
-	double *param, 
-	int *status, 
-	int model_status, 
-	double **fisher,
-	mcmc_data_interface *interface, 
-	void *user_param);
-double RJMCMC_2WF_likelihood_wrapper(
-	double *param, 
-	int *status, 
-	int model_status, 
-	mcmc_data_interface *interface, 
-	void *user_param);
-void pack_local_mod_structure(mcmc_data_interface *interface,
-	double *param,
-	int *status,
-	std::string waveform_extended, 
-	void *parameters, 
-	MCMC_modification_struct *full_struct, 
-	MCMC_modification_struct *local_struct );
-
-void MCMC_fisher_transformations(
-	double *param, 
-	double **fisher, 
-	int dimension,
-	std::string generation_method,
-	bool intrinsic,
-	mcmc_data_interface *interface, 
-	MCMC_modification_struct *mod_struct,
-	void *parameters);
 
 #endif
