+++ tests/src/test_mcmc.cpp
@@ -32,9 +32,6 @@ int stupid_test(int argc, char *argv[]);
 void RT_ERROR_MSG();
 int test_likelihood(int argc, char *argv[]);
 int mcmc_standard_test(int argc, char *argv[]);
-int mcmc_injection(int argc, char *argv[]);
-int mcmc_injection_RJ(int argc, char *argv[]);
-int mcmc_real_data(int argc, char *argv[]);
 int mcmc_rosenbock(int argc, char *argv[]);
 int mcmc_output_class(int argc, char *argv[]);
 int mcmc_RJ_sin(int argc, char *argv[]);
@@ -97,11 +94,13 @@ int main(int argc, char *argv[])
 	}
 	else if(runtime_opt == 1){
 		std::cout<<"MCMC Injected GW"<<std::endl;
-		return mcmc_injection(argc,argv);
+		std::cout<<"Test function deprecated due to function PTMCMC_MH_dynamic_PT_alloc_uncorrelated_GW no longer being in use"<<std::endl;
+		return -1;
 	}
 	else if(runtime_opt == 2){
 		std::cout<<"MCMC data GW"<<std::endl;
-		return mcmc_real_data(argc,argv);
+		std::cout<<"Test Function depcrecated due to function PTMCMC_MH_GW no longer being in use"<<std::endl;
+		return -1;
 	}
 	else if(runtime_opt == 3){
 		std::cout<<"MCMC N-dimensional Rosenbock"<<std::endl;
@@ -117,7 +116,8 @@ int main(int argc, char *argv[])
 	}
 	else if(runtime_opt == 6){
 		std::cout<<"MCMC Injected GW RJ"<<std::endl;
-		return mcmc_injection_RJ(argc,argv);
+		std::cout<<"Test Function deprecated due to function RJPTMCMC_MH_dynamic_PT_alloc_comprehensive_2WF_GW no longer being in use"<<std::endl;
+		return -1;
 	}
 	else if(runtime_opt == 7){
 		std::cout<<"Continue MCMC validation"<<std::endl;
@@ -1294,246 +1294,6 @@ int multiple_continue(int argc, char *argv[])
 	return 0;
 }
 
-int mcmc_injection_RJ(int argc, char *argv[])
-{
-	gen_params injection;
-	injection.mass1 = 36.4;
-	injection.mass2 = 29.3;
-	//injection.mass1 = 24.4;
-	//injection.mass2 = 2.7;
-	//injection.mass1 = 9;
-	//injection.mass2 = 7;
-	double chirpmass = calculate_chirpmass(injection.mass1,injection.mass2);
-	double eta = calculate_eta(injection.mass1,injection.mass2);
-	double q = injection.mass2/injection.mass1;
-	injection.Luminosity_Distance =2000;
-	injection.psi = .2;
-	injection.phiRef = 2.;
-	injection.f_ref = 20.;
-	injection.RA = .275;
-	injection.DEC = -.44;
-	injection.spin1[2] = .03;
-	injection.spin2[2] = .02;
-	injection.spin1[1] = .01;
-	injection.spin2[1] = -.01;
-	injection.spin1[0] = .01;
-	injection.spin2[0] = -.01;
-	injection.incl_angle = .051;
-	//injection.incl_angle = -M_PI +.01;
-	double gps = 1126259462.4;
-	//double gps = 1180922494.5;
-	injection.gmst = gps_to_GMST_radian(gps);
-	injection.equatorial_orientation = false;
-	injection.horizon_coord = false;
-	injection.shift_time = true;
-	injection.shift_phase = true;
-
-	IMRPhenomPv2<double> model;
-	injection.chip = model.PhenomPv2_inplane_spin(&injection);
-	injection.phip = 1;
-
-
-	//injection.Nmod_phi = 1;
-	//injection.delta_phi = new double[1];
-	//injection.delta_phi[0] = 1;
-	//injection.phii = new int[1];
-	//injection.phii[0] = 4;
-	
-	int detect_number = 3;
-	std::string detectors[4] = {"Hanford","Livingston","Virgo","Kagra"};
-	std::string SN[4] = {"AdLIGODesign","AdLIGODesign","AdLIGODesign","KAGRA_pess"};
-	std::string injection_method = "IMRPhenomPv2";
-	double fmin = 20;
-	double fmax =2048;
-	T_mcmc_gw_tool= 4;
-	double tc_ref = T_mcmc_gw_tool*(1-3./4.);
-	//double tc_ref = T_mcmc_gw_tool*(3./4.);
-	double deltaf = 1./T_mcmc_gw_tool;
-	int length = (fmax-fmin)/deltaf;
-	std::complex<double> **data = new std::complex<double>*[detect_number];
-	double **psd = new double*[detect_number];
-	double **freq = new double*[detect_number];
-	double total_snr = 0;
-	double tc=0;
-	injection.tc = 0;
-	for(int i = 0 ; i<detect_number; i++){
-		data[i]= new std::complex<double>[length];
-		psd[i]= new double[length];
-		freq[i]= new double[length];
-		for(int j =0 ; j<length; j++){
-			freq[i][j] = fmin + j*deltaf;
-		}
-		populate_noise(freq[i],SN[i],psd[i],length);
-		for(int j =0 ; j<length; j++){
-			psd[i][j] *= psd[i][j];
-		}
-		double deltat = DTOA_DETECTOR(injection.RA,injection.DEC,injection.gmst,detectors[0],detectors[i]);
-		tc = tc_ref - deltat;
-		tc*=2*M_PI;
-		fourier_detector_response(freq[i],length, data[i],detectors[i],injection_method, &injection, (double *)NULL);
-		for(int j = 0 ; j<length; j++){
-			data[i][j]*=exp(std::complex<double>(0,tc*freq[i][j]));
-		}
-		total_snr += pow_int( calculate_snr_internal(psd[i],data[i],freq[i],length, "SIMPSONS",(double*) NULL, false), 2);
-	}
-	injection.tc=tc_ref;
-	std::cout<<"NETWORK SNR of injection: "<<sqrt(total_snr)<<std::endl;
-
-
-	int data_lengths[detect_number];
-	for(int i = 0 ; i<detect_number; i++){
-		data_lengths[i]=length;	
-	}
-	
-	//#############################################################
-	//#############################################################
-	
-
-	int min_dim = 11;
-	std::string recovery_method_base = "IMRPhenomD";
-	std::string recovery_method_extended = "gIMRPhenomD";
-	int ensemble = 10;
-	//int ensemble = 5;
-	int chains = 40;
-	//int chains = 10;
-	double temps[chains];
-	double c = 1.2;
-	temps[0]=1;
-	for(int i = 1 ; i<chains; i++){
-		temps[i]=temps[i-1]*c;
-	}
-	
-
-
-	//////////////////////////////////////////////
-	//initial_position[0]+=.1;
-	//initial_position[1]+=.4;
-	//initial_position[2]+=.1;
-	//initial_position[3]-=.07;
-	//initial_position[4]+=.7;
-	//initial_position[5]+=.02;
-	//initial_position[6]+=.1;
-	//initial_position[7]+=.4;
-	//initial_position[8]+=.05;
-	//initial_position[9]+=.2;
-	//initial_position[10]+=.1;
-	//initial_position[11]+=.1;
-	//initial_position[12]+=.1;
-	//initial_position[13]+=.1;
-	//initial_position[14]+=.1;
-	//////////////////////////////////////////////
-
-
-	//initial_position[3] = -initial_position[3];
-	//initial_position[1] = 0;
-	//initial_position[0] = 2.;
-	double *seeding = NULL;
-	int swap_freq = 5;
-	int threads = 8;
-	bool pool = true;
-	bool show_progress = true;
-	std::string stat_file = "data/RJ_injection_stat.txt";
-	std::string output_file = "data/RJ_injection_output.csv";
-	std::string ll_file = "data/RJ_injection_ll.csv";
-	std::string checkpoint_file = "data/RJ_injection_checkpoint.csv";
-	std::string ac_file = "";
-	MCMC_modification_struct mod_struct;
-	mod_struct.gIMR_phii = NULL;
-	mod_struct.gIMR_sigmai = NULL;
-	mod_struct.gIMR_betai = NULL;
-	mod_struct.gIMR_alphai = NULL;
-
-	mod_struct.gIMR_Nmod_phi = 3;
-	mod_struct.gIMR_phii = new int[mod_struct.gIMR_Nmod_phi];
-	mod_struct.gIMR_phii[0 ] = 2;
-	mod_struct.gIMR_phii[1 ] = 3;
-	mod_struct.gIMR_phii[2 ] = 4;
-	//mod_struct.gIMR_Nmod_alpha = 1;
-	//mod_struct.gIMR_alphai = new int[mod_struct.gIMR_Nmod_alpha];
-	//mod_struct.gIMR_alphai[0 ] = 2;
-	//mod_struct.gIMR_Nmod_sigma = 2;
-	//mod_struct.gIMR_sigmai = new int[mod_struct.gIMR_Nmod_sigma];
-	//mod_struct.gIMR_sigmai[0 ] = 2;
-	//mod_struct.gIMR_sigmai[1 ] = 3;
-	//mod_struct.gIMR_Nmod_beta = 2;
-	//mod_struct.gIMR_betai = new int[mod_struct.gIMR_Nmod_beta];
-	//mod_struct.gIMR_betai[0 ] = 2;
-	//mod_struct.gIMR_betai[1 ] = 3;
-	
-
-	int max_dim = min_dim + mod_struct.gIMR_Nmod_phi+mod_struct.gIMR_Nmod_sigma+mod_struct.gIMR_Nmod_beta+mod_struct.gIMR_Nmod_alpha;
-
-
-	double spin1sph[3];
-	double spin2sph[3];
-	transform_cart_sph(injection.spin1,spin1sph);
-	transform_cart_sph(injection.spin2,spin2sph);
-	//double initial_position[max_dim]= {injection.RA, sin(injection.DEC),injection.psi, cos(injection.incl_angle), injection.phiRef, T_mcmc_gw_tool-tc_ref, log(injection.Luminosity_Distance),log(chirpmass), eta, spin1sph[0],spin2sph[0],cos(spin1sph[1]),cos(spin2sph[1]),spin1sph[2],spin2sph[2],0,0,0};
-	double initial_position[max_dim]= {injection.RA, sin(injection.DEC),injection.psi, cos(injection.incl_angle), injection.phiRef, T_mcmc_gw_tool-tc_ref, log(injection.Luminosity_Distance),log(chirpmass), eta,injection.spin1[2],injection.spin2[2],0,0,0};
-
-	int initial_status[max_dim];
-	for(int i = 0 ; i<min_dim; i++){
-		initial_status[i] = 1;
-	}
-	for(int i = min_dim ; i<max_dim; i++){
-		initial_status[i] = 1;
-	}
-	write_file("data/injections.csv",initial_position,max_dim);
-
-	int samples = 500;
-	double **output  = allocate_2D_array( samples, max_dim);
-	int **status  = allocate_2D_array_int( samples, max_dim);
-	int *model_status  = NULL;
-	int nested_model_number = 0;
-	int initial_model_status= 0;
-	//if(nested_model_number != 0 ){
-	//	initial_model_status = new int[nested_model_number];
-	//}
-	double t0 = 200;
-	double nu = 10;
-	std::string chain_distribution="double";
-	int max_chunk_size = 5e1;
-	
-	double **prior_ranges = new double*[max_dim-min_dim];
-	for(int i = 0 ; i<max_dim-min_dim; i++){
-		prior_ranges[i] = new double[2];
-		prior_ranges[i][0]=-10;
-		prior_ranges[i][1]=10;
-	}
-	
-	mcmc_sampler_output sampler_output(chains,max_dim);
-	sampler_output.RJ = true;
-	RJPTMCMC_MH_dynamic_PT_alloc_comprehensive_2WF_GW(&sampler_output,output,status,model_status, nested_model_number,max_dim, min_dim , samples, chains, ensemble,initial_position,initial_status, initial_model_status, seeding, (double**)NULL,(int**)NULL,(int*)NULL,prior_ranges,temps, swap_freq, t0,nu,max_chunk_size,chain_distribution,standard_log_prior_D_RJ, threads, pool, show_progress, detect_number, data, psd, freq, data_lengths, gps, detectors, &mod_struct, recovery_method_base, recovery_method_extended, stat_file, output_file, ll_file, checkpoint_file);
-	deallocate_2D_array(output,  samples, max_dim);
-	deallocate_2D_array(status,  samples, max_dim);
-
-	if(mod_struct.gIMR_phii){
-		delete [] mod_struct.gIMR_phii;
-	}
-	if(mod_struct.gIMR_alphai){
-		delete [] mod_struct.gIMR_alphai;
-	}
-	if(mod_struct.gIMR_betai){
-		delete [] mod_struct.gIMR_betai;
-	}
-	if(mod_struct.gIMR_sigmai){
-		delete [] mod_struct.gIMR_sigmai;
-	}
-	
-	for(int i = 0 ; i<detect_number; i++){
-		delete [] freq[i];	
-		delete [] psd[i];	
-		delete [] data[i];	
-	}
-	delete [] freq;
-	delete [] psd;
-	delete [] data;
-	//deallocate_2D_array(jac_spins,  fisher_dim, fisher_dim);
-
-	return 0;
-
-}
-
 double RJ_sin_tilt_range[2];
 int mcmc_RJ_sin(int argc, char *argv[])
 {
@@ -2310,394 +2070,6 @@ double log_rosenbock_prior (double *c,mcmc_data_interface *interface,void *param
 	for(int i = 0 ; i<param->n; i++){
 		if( c[i]<param->bounds_low[i] || c[i] > param->bounds_high[i]){return a;}
 	}
-	return 0;
-}
-int mcmc_real_data(int argc, char *argv[])
-{
-	double gps = 1126259462.4;
-	
-	int detect_number = 2;
-	std::string detectors[2] = {"Hanford","Livingston"};
-	T_mcmc_gw_tool= 32;
-	int dat_length = 65537;
-	double **unpack = allocate_2D_array(dat_length, 4);
-	read_file("data/data_H1.csv", unpack,dat_length , 4);
-	std::complex<double> **data = new std::complex<double>*[detect_number];
-	for(int i = 0 ; i<detect_number; i++){
-		data[i]= new std::complex<double>[dat_length-1];
-	}
-	double **freq = allocate_2D_array( 2,dat_length- 1);
-	double **psd = allocate_2D_array( 2,dat_length- 1);
-	for(int i = 1 ; i < dat_length; i++){
-		data[0][i-1] = std::complex<double>(unpack[i][2],unpack[i][3]);
-		freq[0][i-1] = unpack[i][0];
-		psd[0][i-1] = unpack[i][1];
-	}
-	read_file("data/data_L1.csv", unpack,dat_length , 4);
-	for(int i = 1 ; i < dat_length; i++){
-		data[1][i-1] = std::complex<double>(unpack[i][2],unpack[i][3]);
-		freq[1][i-1] = unpack[i][0];
-		psd[1][i-1] = unpack[i][1];
-	}
-
-	
-
-	int data_lengths[detect_number];
-	for(int i = 0 ; i<detect_number; i++){
-		data_lengths[i]=dat_length-1;	
-	}
-	//#############################################################
-	//#############################################################
-	
-	double **whitened = allocate_2D_array(data_lengths[0],7);
-	for(int i = 0 ; i<data_lengths[0]; i++){
-		whitened[i][0]= freq[0][i];
-		whitened[i][1]=psd[0][i];
-		whitened[i][2]=psd[1][i];
-		whitened[i][3]=real(data[0][i]);
-		whitened[i][4]=imag(data[0][i]);
-		whitened[i][5]=real(data[1][i]);
-		whitened[i][6]=imag(data[1][i]);
-	}	
-	write_file("data/whitened_data.csv",whitened,data_lengths[0],7);
-	deallocate_2D_array(whitened, data_lengths[0],7);
-	
-	//#############################################################
-	//#############################################################
-	
-
-	int dim = 11;
-	std::string recovery_method = "IMRPhenomD";
-	int chains = 20;
-	double temps[chains];
-	double c = 1.2;
-	temps[0]=1;
-	for(int i = 1 ; i<chains; i++){
-		temps[i]=temps[i-1]*c;
-	}
-	
-	double initial_position[dim]= {1.8, sin(-1.2),1, -.9, 1, 32*3./4, log(410),log(32), .24, .0,0};
-	//double initial_position[dim]= {injection.RA, sin(injection.DEC), cos(injection.incl_angle), injection.phiRef, tc_ref, log(injection.Luminosity_Distance),log(chirpmass), eta, injection.spin1[2],injection.spin2[2]};
-	write_file("data/injections.csv",initial_position,dim);
-	//initial_position[3] = -initial_position[3];
-	//initial_position[1] = 0;
-	//initial_position[0] = 2.;
-	double *seeding = NULL;
-	int swap_freq = 3;
-	int threads = 10;
-	bool pool = true;
-	bool show_progress = true;
-	std::string stat_file = "data/experiment_stat.txt";
-	std::string output_file = "data/experiment_output.csv";
-	std::string ll_file = "data/experiment_ll.csv";
-	std::string checkpoint_file = "data/experiment_checkpoint.csv";
-	//std::string ac_file = "data/injection_ac.csv";
-	std::string ac_file = "";
-	MCMC_modification_struct mod_struct;
-	
-	int samples = 50000;
-	double ***output  = allocate_3D_array(chains, samples, dim);
-	PTMCMC_MH_GW(output, dim , samples, chains, initial_position, seeding, (double**)NULL,temps, swap_freq, standard_log_prior_D, threads, pool, show_progress, detect_number, data, psd, freq, data_lengths, gps, detectors, &mod_struct, recovery_method, stat_file, output_file, ac_file,ll_file, checkpoint_file);
-	deallocate_3D_array(output, chains, samples, dim);
-	//int samples = 2000;
-	//double **output  = allocate_2D_array( samples, dim);
-	//double t0 = 1000;
-	//double nu = 500;
-	//double corr_threshold = .01;
-	//double corr_segments = 10;
-	//double corr_converge_thresh = .1;
-	//double corr_target_ac = .01;
-	//std::string chain_distribution="double";
-	//PTMCMC_MH_dynamic_PT_alloc_uncorrelated_GW(output, dim , samples, chains, chains,initial_position, seeding, temps, swap_freq, t0,nu,corr_threshold, corr_segments, corr_converge_thresh, corr_target_ac,chain_distribution,standard_log_prior_D, threads, pool, show_progress, detect_number, data, psd, freq, data_lengths, gps, detectors, Nmod, bppe, recovery_method, stat_file, output_file, ll_file, checkpoint_file);
-	//deallocate_2D_array(output,  samples, dim);
-
-	
-
-	return 0;
-}
-int mcmc_injection(int argc, char *argv[])
-{
-	bool GLQ = false;
-	gen_params injection;
-	injection.mass1 = 36.4;
-	injection.mass2 = 29.3;
-	double chirpmass = calculate_chirpmass(injection.mass1,injection.mass2);
-	double eta = calculate_eta(injection.mass1,injection.mass2);
-	double q = injection.mass2/injection.mass1;
-	injection.Luminosity_Distance =500;
-	injection.psi = .2;
-	injection.phiRef = 2.;
-	injection.f_ref = 20.;
-	injection.RA = .275;
-	injection.DEC = -.44;
-	injection.spin1[2] = .3;
-	injection.spin2[2] = .2;
-	injection.spin1[1] = .5;
-	injection.spin2[1] = -.5;
-	injection.spin1[0] = .01;
-	injection.spin2[0] = -.01;
-	injection.incl_angle = .51;
-	double gps = 1126259462.4;
-	injection.gmst = gps_to_GMST_radian(gps);
-	std::cout<<"GMST: "<<injection.gmst<<std::endl;
-	injection.equatorial_orientation = false;
-	injection.horizon_coord = false;
-	injection.shift_time = true;
-	injection.shift_phase = true;
-
-	IMRPhenomPv2<double> model;
-	injection.chip = model.PhenomPv2_inplane_spin(&injection);
-	injection.phip = 1;
-
-
-	injection.Nmod = 1;
-	injection.bppe = new double[1];
-	injection.betappe = new double[1];
-	injection.bppe[0] = -1;
-	injection.betappe[0] = 0;
-	
-	int detect_number = 4;
-	std::string detectors[4] = {"Hanford","Livingston","Virgo","Kagra"};
-	std::string SN[4] = {"AdLIGODesign","AdLIGODesign","AdLIGODesign","KAGRA_pess"};
-	std::string injection_method = "ppE_IMRPhenomD_Inspiral";
-	double fmin = 10;
-	double fmax =512;
-	T_mcmc_gw_tool= 4;
-	int length = 10;
-	int fish_length = 1000;
-	double **psd;
-	double **freq;
-	double **weights;
-	double **fish_freq;
-	double **fish_psd;
-	std::complex<double> **data ;
-	double tc_ref=0;
-	if(GLQ){
-		std::cout<<"Using GLQ"<<std::endl;
-		tc_ref = T_mcmc_gw_tool*(1-3./4.);
-		//double tc_ref = T_mcmc_gw_tool*(3./4.);
-		length = 1000;
-		data = new std::complex<double>*[detect_number];
-		psd = new double*[detect_number];
-		fish_psd = new double*[detect_number];
-		fish_freq = new double*[detect_number];
-		freq = new double*[detect_number];
-		weights = new double*[detect_number];
-		double total_snr = 0;
-		double tc=0;
-		injection.tc = 0;
-		for(int i = 0 ; i<detect_number; i++){
-			data[i]= new std::complex<double>[length];
-			psd[i]= new double[length];
-			freq[i]= new double[length];
-			fish_psd[i]= new double[fish_length];
-			fish_freq[i]= new double[fish_length];
-			weights[i]= new double[length];
-			gauleg(log10(fmin),log10(fmax),freq[i],weights[i],length);
-			for(int j =0 ; j<length; j++){
-				freq[i][j] = pow(10,freq[i][j]);
-			}
-			for(int j =0 ; j<fish_length; j++){
-				fish_freq[i][j] = fmin+ (fmax-fmin)/fish_length;
-			}
-			populate_noise(freq[i],SN[i],psd[i],length);
-			populate_noise(fish_freq[i],SN[i],fish_psd[i],fish_length);
-			for(int j =0 ; j<length; j++){
-				psd[i][j] *= psd[i][j];
-			}
-			for(int j =0 ; j<fish_length; j++){
-				fish_psd[i][j] *= fish_psd[i][j];
-			}
-			double deltat = DTOA_DETECTOR(injection.RA,injection.DEC,injection.gmst,detectors[0],detectors[i]);
-			//injection.tc = tc_ref + deltat;
-			tc = tc_ref - deltat;
-			tc*=2*M_PI;
-			fourier_detector_response(freq[i],length, data[i],detectors[i],injection_method, &injection, (double *)NULL);
-			for(int j = 0 ; j<length; j++){
-				data[i][j]*=exp(std::complex<double>(0,tc*freq[i][j]));
-			}
-			total_snr += pow_int( calculate_snr_internal(psd[i],data[i],freq[i],length, "GAUSSLEG",weights[i], true), 2);
-		}
-		injection.tc=tc_ref;
-		std::cout<<"NETWORK SNR of injection: "<<sqrt(total_snr)<<std::endl;
-	}
-	else{
-		tc_ref = T_mcmc_gw_tool*(1-3./4.);
-		//double tc_ref = T_mcmc_gw_tool*(3./4.);
-		double deltaf = 1./T_mcmc_gw_tool;
-		length = (fmax-fmin)/deltaf;
-		data = new std::complex<double>*[detect_number];
-		psd = new double*[detect_number];
-		freq = new double*[detect_number];
-		double total_snr = 0;
-		double tc=0;
-		injection.tc = 0;
-		for(int i = 0 ; i<detect_number; i++){
-			data[i]= new std::complex<double>[length];
-			psd[i]= new double[length];
-			freq[i]= new double[length];
-			for(int j =0 ; j<length; j++){
-				freq[i][j] = fmin + j*deltaf;
-			}
-			populate_noise(freq[i],SN[i],psd[i],length);
-			for(int j =0 ; j<length; j++){
-				psd[i][j] *= psd[i][j];
-			}
-			double deltat = DTOA_DETECTOR(injection.RA,injection.DEC,injection.gmst,detectors[0],detectors[i]);
-			//injection.tc = tc_ref + deltat;
-			tc = tc_ref - deltat;
-			tc*=2*M_PI;
-			fourier_detector_response(freq[i],length, data[i],detectors[i],injection_method, &injection, (double *)NULL);
-			for(int j = 0 ; j<length; j++){
-				data[i][j]*=exp(std::complex<double>(0,tc*freq[i][j]));
-			}
-			total_snr += pow_int( calculate_snr_internal(psd[i],data[i],freq[i],length, "SIMPSONS",(double*) NULL, false), 2);
-		}
-		injection.tc=tc_ref;
-		std::cout<<"NETWORK SNR of injection: "<<sqrt(total_snr)<<std::endl;
-	}
-
-	int data_lengths[detect_number];
-	int fisher_data_lengths[detect_number];
-	for(int i = 0 ; i<detect_number; i++){
-		data_lengths[i]=length;	
-		fisher_data_lengths[i] = fish_length;
-	}
-	//###################################################################
-	//###################################################################
-	double **weights_temp= new double*[detect_number];
-	injection.tc = T_mcmc_gw_tool - injection.tc ;
-	double sample_ll = MCMC_likelihood_extrinsic(true, &injection, injection_method, data_lengths,freq, data,psd,  weights_temp, "SIMPSONS",false, detectors, detect_number);
-	std::cout<<"Sample LL: "<<sample_ll<<std::endl;
-	delete [] weights_temp;
-	//###################################################################
-	//###################################################################
-	
-	injection.shift_time = false;
-	injection.shift_phase = false;
-	double fisher_dim = 12;
-	double **temp_fisher = allocate_2D_array(fisher_dim,fisher_dim);
-	double **fisher = allocate_2D_array(fisher_dim,fisher_dim);
-	double **cov = allocate_2D_array(fisher_dim,fisher_dim);
-	for(int i = 0 ; i<fisher_dim; i++){
-		for(int j = 0 ; j<fisher_dim; j++){
-			fisher[i][j]=0;	
-		}
-	}
-	for(int i = 0 ; i<detect_number; i++){
-		if(GLQ){
-			fisher_autodiff(freq[i],length, "ppE_IMRPhenomD_Inspiral", detectors[i],detectors[0],temp_fisher,fisher_dim,&injection, "GAUSSLEG",weights[i],true, psd[i],NULL,NULL);
-		}
-		else{
-			fisher_autodiff(freq[i],length, "ppE_IMRPhenomD_Inspiral", detectors[i],detectors[0],temp_fisher,fisher_dim,&injection, "SIMPSONS",(double *)NULL,false, psd[i],NULL,NULL);
-		}
-		for(int k = 0 ; k<fisher_dim; k++){
-			for(int j = 0 ; j<fisher_dim; j++){
-				fisher[k][j]+=temp_fisher[k][j];	
-			}
-		}
-	}
-	gsl_LU_matrix_invert(fisher, cov, fisher_dim);
-	std::cout<<"Fisher estimates of covariances: "<<std::endl;
-	for(int i = 0  ;i<fisher_dim; i++){
-		std::cout<<i<<" "<<sqrt(cov[i][i])<<" "<<std::endl;
-	}
-	std::cout<<std::endl;
-	deallocate_2D_array(temp_fisher,fisher_dim,fisher_dim);
-	deallocate_2D_array(fisher,fisher_dim,fisher_dim);
-	deallocate_2D_array(cov,fisher_dim,fisher_dim);
-	injection.shift_time = true;
-	injection.shift_phase = true;
-	
-
-	int dim = 12;
-	std::string recovery_method = "ppE_IMRPhenomD_Inspiral";
-	int ensemble = 15;
-	//int ensemble = 5;
-	int chains = 405;
-	//int chains = 10;
-	double temps[chains];
-	double c = 1.2;
-	temps[0]=1;
-	for(int i = 1 ; i<chains; i++){
-		temps[i]=temps[i-1]*c;
-	}
-	
-	double spin1sph[3];
-	double spin2sph[3];
-	transform_cart_sph(injection.spin1,spin1sph);
-	transform_cart_sph(injection.spin2,spin2sph);
-	//double initial_position[dim]= {injection.RA, sin(injection.DEC),injection.psi, cos(injection.incl_angle), injection.phiRef, T_mcmc_gw_tool-tc_ref, log(injection.Luminosity_Distance),log(chirpmass), eta, spin1sph[0],spin2sph[0],cos(spin1sph[1]),cos(spin2sph[1]),spin1sph[2],spin2sph[2],injection.betappe[0]};
-	double initial_position[dim]= {injection.RA, sin(injection.DEC),injection.psi, cos(injection.incl_angle), injection.phiRef, T_mcmc_gw_tool-tc_ref, log(injection.Luminosity_Distance),log(chirpmass), eta, injection.spin1[2],injection.spin2[2],0};
-	//double initial_position[dim]= {log(chirpmass), eta, injection.spin1[2],injection.spin2[2],0};
-	//double initial_position[dim]= {injection.RA, sin(injection.DEC),injection.psi, cos(injection.incl_angle), injection.phiRef, T_mcmc_gw_tool-tc_ref, log(injection.Luminosity_Distance),chirpmass, q, injection.spin1[2],injection.spin2[2]};
-	//double initial_position[dim]= {injection.RA, sin(injection.DEC),injection.psi, cos(injection.incl_angle), injection.phiRef, T_mcmc_gw_tool-tc_ref, log(injection.Luminosity_Distance),injection.mass1, injection.mass2, injection.spin1[2],injection.spin2[2]};
-	//double initial_position[dim]= {injection.RA, sin(injection.DEC),injection.psi, cos(injection.incl_angle), injection.phiRef, tc_ref, log(injection.Luminosity_Distance),log(chirpmass), eta, injection.spin1[2],injection.spin2[2]};
-	//double initial_position[dim]= {injection.RA, sin(injection.DEC), cos(injection.incl_angle), injection.phiRef, tc_ref, log(injection.Luminosity_Distance),log(chirpmass), eta, injection.spin1[2],injection.spin2[2]};
-	write_file("data/injections.csv",initial_position,dim);
-
-
-
-
-	double *seeding = NULL;
-	int swap_freq = 5;
-	int threads = 8;
-	bool pool = true;
-	bool show_progress = true;
-	std::string stat_file = "data/injection_stat.txt";
-	std::string output_file = "data/injection_output.csv";
-	std::string ll_file = "data/injection_ll.csv";
-	std::string checkpoint_file = "data/injection_checkpoint.csv";
-	//std::string ac_file = "data/injection_ac.csv";
-	std::string ac_file = "";
-	MCMC_modification_struct mod_struct;
-	mod_struct.ppE_Nmod = 1;
-	double b = -1;
-	mod_struct.bppe =&b ;
-	
-	if(GLQ){
-		mod_struct.weights=weights;
-		mod_struct.log10F=true;
-		mod_struct.fisher_log10F=false;
-		mod_struct.fisher_GAUSS_QUAD=false;
-		mod_struct.GAUSS_QUAD=true;
-		mod_struct.fisher_freq = fish_freq;
-		mod_struct.fisher_PSD= fish_psd;
-		mod_struct.fisher_length= fisher_data_lengths;
-	}
-	
-	int samples = 1000;
-	double **output  = allocate_2D_array( samples, dim);
-	double t0 = 1000;
-	double nu = 100;
-	std::string chain_distribution="double";
-	int max_chunk_size = 1e6;
-	mcmc_sampler_output sampler_output(chains,dim);
-	PTMCMC_MH_dynamic_PT_alloc_uncorrelated_GW(&sampler_output,output, dim , samples, chains, ensemble,initial_position, seeding, (double**)NULL,temps, swap_freq, t0,nu,max_chunk_size,chain_distribution,standard_log_prior_D, threads, pool, show_progress, detect_number, data, psd, freq, data_lengths, gps, detectors, &mod_struct, recovery_method, stat_file, output_file, ll_file, checkpoint_file);
-	deallocate_2D_array(output,  samples, dim);
-
-	
-	for(int i = 0 ; i<detect_number; i++){
-		delete [] freq[i];	
-		delete [] psd[i];	
-		delete [] data[i];	
-		if(GLQ){
-			delete [] weights[i];
-			delete [] fish_freq[i];
-			delete [] fish_psd[i];
-		}
-	}
-	delete [] injection.bppe;
-	delete [] injection.betappe;
-	delete [] freq;
-	delete [] psd;
-	delete [] data;
-	if(GLQ){
-		delete [] weights;
-		delete [] fish_freq;
-		delete [] fish_psd;
-	}
-
 	return 0;
 }
 int mcmc_standard_test(int argc, char *argv[])
